Change time format from uint40 to uint32
Because it's enough.
uint32 has a capable of handling time to 2106.
Small storage size is advantage.

---

Original commit (chain-contracts): 8069c0ad82ea45e584d4053169ab4221a0f35ea0

+++ chain-contracts/contracts/Web.sol
@@ -19,10 +19,10 @@ contract Web is UserGoldControllable, UserNormalPlanetControllable {
     view
     returns (
     uint200 confirmedGold,
-    uint40 goldConfirmedAt,
+    uint32 goldConfirmedAt,
     uint16[] unpIds, // [id, normalPlanetId, ...]
     uint8[] unpRanks,
-    uint40[] unpTimes, // [rankupedAt, createdAt, ...]
+    uint32[] unpTimes, // [rankupedAt, createdAt, ...]
     int16[] unpAxialCoordinates // [q, r, ...]
   )
   {
@@ -35,7 +35,7 @@ contract Web is UserGoldControllable, UserNormalPlanetControllable {
 
     unpIds = new uint16[](userPlanetsCount * 2);
     unpRanks = new uint8[](userPlanetsCount);
-    unpTimes = new uint40[](userPlanetsCount * 2);
+    unpTimes = new uint32[](userPlanetsCount * 2);
     unpAxialCoordinates = new int16[](userPlanetsCount * 2);
     uint counter = 0;
